增加对于exception的返回json串的公共方法

+++ src/main/java/com/chings/core/service/impl/ArticleServiceImpl.java
@@ -111,8 +111,10 @@ public class ArticleServiceImpl implements IArticleService {
 		return new ArrayList<Map>();
 	}
 
+
 	@Override
-	public int deleteArticle(int id) {
+	public long deleteArticle(long id) {
 		return articleDao.deleteArticle(id);
 	}
+
 }
